Fixed another bunch of compilier warnings.
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@8405 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/plugins/fileloader/scribus134format/scribus134format.cpp
@@ -2093,14 +2093,14 @@ PageItem* Scribus134Format::PasteItem(QDomElement *obj, ScribusDoc *doc)
 	}
 	currItem->FrameType = obj->attribute("FRTYPE", "0").toInt();
 	int startArrowIndex = obj->attribute("startArrowIndex", "0").toInt();
-	if (startArrowIndex < 0 || startArrowIndex > doc->arrowStyles.size())
+	if ((startArrowIndex < 0) || (startArrowIndex > static_cast<int>(doc->arrowStyles.size())))
 	{
 		qDebug(QString("scribus134format: invalid arrow index: %").arg(startArrowIndex));
 		startArrowIndex = 0;
 	}
 	currItem->setStartArrowIndex(startArrowIndex);
 	int endArrowIndex = obj->attribute("endArrowIndex", "0").toInt();
-	if (endArrowIndex < 0 || endArrowIndex > doc->arrowStyles.size())
+	if ((endArrowIndex < 0) || (endArrowIndex > static_cast<int>(doc->arrowStyles.size())))
 	{
 		qDebug(QString("scribus134format: invalid arrow index: %").arg(endArrowIndex));
 		endArrowIndex = 0;
